修复了许多bug

+++ src/main/java/com/syes/syes_springboot/entity/Auction.java
@@ -32,7 +32,7 @@ public class Auction implements Serializable {
 
     private Integer itemid;
 
-    private Integer price;
+    private float price;
 
     private LocalDateTime time;
